+++ packages/web/src/__tests__/orchestrators.test.tsx
@@ -1,16 +1,14 @@
 import { describe, it, expect, vi, beforeEach } from "vitest";
-import { render, screen, fireEvent, waitFor } from "@testing-library/react";
-import { OrchestratorSelector } from "@/components/OrchestratorSelector";
+import { render, screen } from "@testing-library/react";
 import OrchestratorsRoute from "@/app/orchestrators/page";
 import { getServices } from "@/lib/services";
 import { getAllProjects } from "@/lib/project-name";
 
 // ── Mocks ─────────────────────────────────────────────────────────────
 
-const mockPush = vi.fn();
 vi.mock("next/navigation", () => ({
   useRouter: () => ({
-    push: mockPush,
+    push: vi.fn(),
   }),
 }));
 
@@ -31,123 +29,6 @@ vi.mock("@/lib/project-name", () => ({
 global.fetch = vi.fn();
 
 // ── Tests ─────────────────────────────────────────────────────────────
-
-describe("OrchestratorSelector Component", () => {
-  const defaultProps = {
-    orchestrators: [
-      {
-        id: "orch-1",
-        projectId: "my-app",
-        projectName: "My App",
-        status: "working",
-        activity: "active",
-        createdAt: new Date(Date.now() - 3600000 * 2).toISOString(), // 2h ago
-        lastActivityAt: new Date(Date.now() - 60000 * 5).toISOString(), // 5m ago
-      },
-    ],
-    projectId: "my-app",
-    projectName: "My App",
-    projects: [{ id: "my-app", name: "My App" }],
-    error: null,
-  };
-
-  beforeEach(() => {
-    vi.clearAllMocks();
-  });
-
-  it("renders orchestrators and handles spawn success", async () => {
-    (global.fetch as any).mockResolvedValue({
-      ok: true,
-      json: async () => ({ orchestrator: { id: "new-orch" } }),
-    });
-
-    render(<OrchestratorSelector {...defaultProps} />);
-
-    expect(screen.getByText("orch-1")).toBeInTheDocument();
-    expect(screen.getByText(/2h ago/)).toBeInTheDocument();
-    expect(screen.getByText(/5m ago/)).toBeInTheDocument();
-
-    const spawnBtn = screen.getByText("Start New Orchestrator");
-    fireEvent.click(spawnBtn);
-
-    await waitFor(() => {
-      expect(mockPush).toHaveBeenCalledWith("/sessions/new-orch");
-    });
-  });
-
-  it("covers relative time for days and status colors/labels", () => {
-    const wideProps = {
-      ...defaultProps,
-      orchestrators: [
-        {
-          id: "orch-2",
-          projectId: "my-app",
-          projectName: "My App",
-          status: "ci_failed",
-          activity: "waiting_input",
-          createdAt: new Date(Date.now() - 3600000 * 50).toISOString(), // 2d ago
-          lastActivityAt: null,
-        },
-        {
-          id: "orch-3",
-          projectId: "my-app",
-          projectName: "My App",
-          status: "killed",
-          activity: "ready",
-          createdAt: new Date(Date.now() - 1000).toISOString(), // Just now
-          lastActivityAt: null,
-        },
-        {
-           id: "orch-4",
-           projectId: "my-app",
-           projectName: "My App",
-           status: "unknown",
-           activity: "blocked",
-           createdAt: new Date().toISOString(),
-           lastActivityAt: null,
-        },
-        {
-           id: "orch-5",
-           projectId: "my-app",
-           projectName: "My App",
-           status: "mergeable",
-           activity: "exited",
-           createdAt: new Date().toISOString(),
-           lastActivityAt: null,
-        }
-      ],
-    };
-
-    render(<OrchestratorSelector {...wideProps} />);
-    expect(screen.getByText(/2d ago/)).toBeInTheDocument();
-    expect(screen.getAllByText(/Just now/i).length).toBeGreaterThan(0);
-    expect(screen.getByText(/Waiting/)).toBeInTheDocument();
-    expect(screen.getByText(/Ready/)).toBeInTheDocument();
-    expect(screen.getByText(/Blocked/)).toBeInTheDocument();
-    expect(screen.getByText(/Exited/)).toBeInTheDocument();
-    expect(screen.getByText(/ci failed/i)).toBeInTheDocument();
-  });
-
-  it("handles spawn failure with error message", async () => {
-    (global.fetch as any).mockResolvedValue({
-      ok: false,
-      json: async () => ({ error: "Server error" }),
-    });
-
-    render(<OrchestratorSelector {...defaultProps} orchestrators={[]} />);
-    fireEvent.click(screen.getByText("Start New Orchestrator"));
-
-    await waitFor(() => {
-      expect(screen.getByText("Server error")).toBeInTheDocument();
-    });
-  });
-
-  it("renders error state from props", () => {
-    render(<OrchestratorSelector {...defaultProps} error="Project not found" />);
-    expect(screen.getByText("Project not found")).toBeInTheDocument();
-    expect(screen.getByText("Go to Dashboard")).toHaveAttribute("href", "/");
-  });
-});
 
 describe("Orchestrators Page (OrchestratorsRoute)", () => {
   beforeEach(() => {
